+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow axis label gap fix for numberical axis type.
 ### Addded
 - Color palette chooser.
+- Option for dropdown menus to update the title of the dropdown tab when selecting an item.
 - Multivariate histograms.
 
 ## [v0.3.1]